caiman >1.11, python 3.9, suite2p not working above py3.9

+++ studio/app/optinist/wrappers/caiman/motion_correction.py
@@ -61,7 +61,7 @@ def caiman_mc(
         .transpose(2, 0, 1)
     )
 
-    rois = rois.astype(np.float)
+    rois = rois.astype(np.float32)
 
     for i, _ in enumerate(rois):
         rois[i] *= i + 1
